<update> read write data nio

+++ game-server/src/main/java/nro/server/network/nro/client_packets/AClientPacketHandler.java
@@ -7,7 +7,7 @@ import nro.server.network.nro.NroConnection.State;
 @Retention(RetentionPolicy.RUNTIME)
 public @interface AClientPacketHandler {
 
-    int command();
+    byte command();
     State[] validStates() default {State.CONNECTED};
 
 }
